Renamed Contributor as User.
