+++ rad2/src/commands/project.rs
@@ -27,7 +27,6 @@ use librad::{
     git::GitProject,
     keys::device,
     meta,
-    meta::contributor::ContributorData,
     paths::Paths,
     project::{Project, ProjectId},
 };
@@ -131,7 +130,7 @@ fn init_project(
         file_name.map(|f| f.to_string_lossy().to_string()).unwrap()
     });
 
-    let contrib_meta = ContributorData::new().set_profile(profile).build()?;
+    let contrib_meta = meta::UserData::new().set_profile(profile).build()?;
 
     let proj =
         GitProject::builder(&project_name, &key, contrib_meta).init_project(paths, &sources)?;
